// frontend/src/api/__generated__/models/KeywordCategoryResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 4차 프로젝트 6팀 March API
 * 6팀 March의 백엔드 API 명세서입니다.
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { KeywordResponse } from './KeywordResponse';
import {
    KeywordResponseFromJSON,
    KeywordResponseFromJSONTyped,
    KeywordResponseToJSON,
    KeywordResponseToJSONTyped,
} from './KeywordResponse';

/**
 * 
 * @export
 * @interface KeywordCategoryResponse
 */
export interface KeywordCategoryResponse {
    /**
     * 
     * @type {number}
     * @memberof KeywordCategoryResponse
     */
    categoryId?: number;
    /**
     * 
     * @type {string}
     * @memberof KeywordCategoryResponse
     */
    categoryName?: string;
    /**
     * 
     * @type {boolean}
     * @memberof KeywordCategoryResponse
     */
    multipleChoice?: boolean;
    /**
     * 
     * @type {Array<KeywordResponse>}
     * @memberof KeywordCategoryResponse
     */
    keywords?: Array<KeywordResponse>;
}

/**
 * Check if a given object implements the KeywordCategoryResponse interface.
 */
export function instanceOfKeywordCategoryResponse(value: object): value is KeywordCategoryResponse {
    return true;
}

export function KeywordCategoryResponseFromJSON(json: any): KeywordCategoryResponse {
    return KeywordCategoryResponseFromJSONTyped(json, false);
}

export function KeywordCategoryResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): KeywordCategoryResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'categoryId': json['categoryId'] == null ? undefined : json['categoryId'],
        'categoryName': json['categoryName'] == null ? undefined : json['categoryName'],
        'multipleChoice': json['multipleChoice'] == null ? undefined : json['multipleChoice'],
        'keywords': json['keywords'] == null ? undefined : ((json['keywords'] as Array<any>).map(KeywordResponseFromJSON)),
    };
}

export function KeywordCategoryResponseToJSON(json: any): KeywordCategoryResponse {
    return KeywordCategoryResponseToJSONTyped(json, false);
}

export function KeywordCategoryResponseToJSONTyped(value?: KeywordCategoryResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'categoryId': value['categoryId'],
        'categoryName': value['categoryName'],
        'multipleChoice': value['multipleChoice'],
        'keywords': value['keywords'] == null ? undefined : ((value['keywords'] as Array<any>).map(KeywordResponseToJSON)),
    };
}